% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_col_eta2}
\alias{ca_col_eta2}
\title{Compute col eta2}
\usage{
ca_col_eta2(contrib, X, eigs)
}
\arguments{
\item{contrib}{contribution matrix}

\item{X}{contribution matrix}

\item{eigs}{eigs computed by \code{ca_weighted_eigen}}
}
\description{
Return col eta2 for each correspondence component
}
\details{
\code{ca_col_eta2()} is only used by \code{facto_mca()} to compute eta2 for columns.
}
