% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_row_contrib}
\alias{ca_row_contrib}
\title{Compute row contributions}
\usage{
ca_row_contrib(
  row_coords,
  X,
  weighted_row = rowSums(X),
  weighted_col = colSums(X)
)
}
\arguments{
\item{row_coords}{row coordinates}

\item{X}{standardized matrix}

\item{weighted_row}{row weights}

\item{weighted_col}{column weights}
}
\description{
Return row contributions for each correspondence component
}
\examples{
library(FactoMineR2)

X_scaled <- mtcars[,c(2,8:11)] |>
  ca_standardize()

X_scaled |>
  ca_weighted_eigen() |>
  ca_row_coords() |>
  ca_row_contrib(X_scaled) |>
  head()
}
