% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_row_cos2}
\alias{ca_row_cos2}
\title{Compute row squared cosines}
\usage{
ca_row_cos2(row_coords, X)
}
\arguments{
\item{row_coords}{row coordinates}

\item{X}{standardized matrix}
}
\description{
Return row squared cosines for each correspondence component
}
\examples{
library(FactoMineR2)

X_scaled <- mtcars[,c(2,8:11)] |>
  ca_standardize()

X_scaled |>
  ca_weighted_eigen() |>
  ca_row_coords() |>
  ca_row_cos2(X_scaled) |>
  head()
}
