% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_row_inertia}
\alias{ca_row_inertia}
\title{Compute row contributions}
\usage{
ca_row_inertia(X)
}
\arguments{
\item{X}{standardized matrix}
}
\description{
Return row contributions for each correspondence component
}
\examples{
library(FactoMineR2)

mtcars[,c(2,8:11)] |>
  ca_standardize() |>
  ca_row_inertia()
}
