% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_eig.R
\name{ca_weighted_eigen}
\alias{ca_weighted_eigen}
\title{Compute eigvalues and eigvectors for CA}
\usage{
ca_weighted_eigen(X, weighted_row = rowSums(X), weighted_col = colSums(X))
}
\arguments{
\item{X}{X_active}

\item{weighted_row}{row weights}

\item{weighted_col}{column weights}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\examples{
library(FactoMineR2)

iris[,-5] |>
  ca_standardize() |>
  ca_weighted_eigen()
}
