% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{eigvalues}
\alias{eigvalues}
\title{Compute eigvalues}
\usage{
eigvalues(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigvalues of a matrix
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize(type = "norm") |>
  eigvalues()

}
