% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{eigvectors}
\alias{eigvectors}
\title{Compute eigvectors}
\usage{
eigvectors(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigvectors of a matrix
}
\examples{
library(FactoMineR2)

iris[,-5] |>
 standardize(type = "norm") |>
 eigvectors()

}
