% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{get_eig}
\alias{get_eig}
\title{Compute eigvalues and eigvectors}
\usage{
get_eig(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\examples{
library(FactoMineR2)

iris[,-5] |>
 standardize(type = "norm") |>
 get_eig()

}
