% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoding.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{Perform One hot encoding}
\usage{
one_hot_encoding(M)
}
\arguments{
\item{M}{a matrix}
}
\description{
Return a matrix with one hot encoding
}
\examples{
library(FactoMineR2)

new_matrix <- one_hot_encoding(iris)
}
