% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_contrib}
\alias{pca_ind_contrib}
\title{Compute individual contributions}
\usage{
pca_ind_contrib(ind_coords)
}
\arguments{
\item{ind_coords}{individual coordinates}
}
\description{
Return indivdual contributions for each principal component
}
\examples{
library(FactoMineR2)

X_active <- iris[,-5] |>
  standardize(type = "norm")

vectors <- X_active |>
  eigvectors()

X_active |>
  pca_ind_coords(vectors) |>
  pca_ind_contrib()

}
