% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_coords}
\alias{pca_ind_coords}
\title{Compute coordinates for individuals}
\usage{
pca_ind_coords(eigs)
}
\arguments{
\item{eigs}{eigs computed by \code{get_eigen} or \code{get_weighted_eigen}}
}
\description{
Return principal component for individuals
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  get_weighted_eigen() |>
  pca_ind_coords() |>
  head()
}
