% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_coords}
\alias{pca_var_coords}
\alias{pca_var_cor}
\title{Compute variable coordinates}
\usage{
pca_var_coords(eigenvalues, eigenvectors)

pca_var_cor(eigenvalues, eigenvectors)
}
\arguments{
\item{eigenvalues}{eigenvalues}

\item{eigenvectors}{eigenvectors}
}
\description{
Return variable coordinates
}
\examples{
library(FactoMineR2)

X_active <- iris[, -5] |>
  standardize_norm()

vectors <- X_active |>
  eigvectors()

X_active |>
  eigvalues() |>
  pca_var_coords(vectors)

}
