% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_cor}
\alias{pca_var_cor}
\title{Compute variable correlation}
\usage{
pca_var_cor(eigs)
}
\arguments{
\item{eigs}{eigs computed by \code{get_eigen} or \code{get_weighted_eigen}}
}
\description{
Return variable correlation
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  get_weighted_eigen() |>
  pca_var_cor() |>
  head()
}
