% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_cos2}
\alias{pca_var_cos2}
\title{Compute variable squared cosines}
\usage{
pca_var_cos2(var_coords)
}
\arguments{
\item{var_coords}{variable coordinates}
}
\description{
Return variable squared cosines
}
\examples{
library(FactoMineR2)

X_active <- iris[, -5] |>
  standardize(type = "norm")

vectors <- X_active |>
  eigvectors()

X_active |>
  eigvalues() |>
  pca_var_coords(vectors) |>
  pca_var_cos2()

}
