% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Data standardization}
\usage{
standardize(X, type = c("FactoMineR", "norm"), ...)
}
\arguments{
\item{X}{matrix}

\item{type}{nature of standardization}

\item{...}{additional arguments}
}
\description{
Perform data standardization for multivariate exploratory data analysis.
}
\details{
Standardization depends on what you need to perform factor analysis. Actually,
there are only two types:

* \code{type = "FactoMineR"}: standardization is performed by centering the
data matrix and dividing by the square root of the sum of squares of the
weights. This is the default standardization in FactoMineR::PCA(..., scale.unit = TRUE).

* \code{type = "norm"}: standardization is performed by centering and scaling
the data matrix. This is the default standardization in R.
}
\examples{
library(FactoMineR2)

iris[,-5] |>
 standardize(type = "norm")

}
