% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_col_inertia}
\alias{ca_col_inertia}
\title{Compute col inertia}
\usage{
ca_col_inertia(X)
}
\arguments{
\item{X}{standardized matrix}
}
\description{
Return col inertia for each correspondence component
}
\examples{
library(FactoMineR2)

mtcars[, c(2, 8:11)] |>
  ca_standardize() |>
  ca_col_inertia()
}
