% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_eig.R
\name{ca_weighted_eigen}
\alias{ca_weighted_eigen}
\title{Compute eigenvalues and eigenvectors for CA}
\usage{
ca_weighted_eigen(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigenvalues and eigenvectors of a matrix
}
\examples{
library(booklet)

mtcars[, c(2, 8:11)] |>
  ca_standardize() |>
  ca_weighted_eigen() |>
  head()
}
