% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facto_mca.R
\name{facto_mca}
\alias{facto_mca}
\title{Perform MCA with FactoMineR's style}
\usage{
facto_mca(X, ncp = 5, row_sup = NULL, col_sup = NULL, weighted_row = NULL)
}
\arguments{
\item{X}{a data frame with n rows (individuals) and p columns (numeric variables)}

\item{ncp}{an integer, the number of components to keep (value set by default)}

\item{row_sup}{supplementary rows}

\item{col_sup}{supplementary columns}

\item{weighted_row}{row weights}
}
\description{
Return MCA results with FactoMineR's style
}
\examples{
library(FactoMineR2)

res <- facto_mca(X = mtcars[, c(2, 8:11)], ncp = 2)
}
