% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facto_pca.R
\name{facto_pca}
\alias{facto_pca}
\title{Perform PCA with FactoMineR's style}
\usage{
facto_pca(X, ncp = 5, scale.unit = TRUE, ind_sup = NULL, quanti_sup = NULL)
}
\arguments{
\item{X}{a data frame with n rows (individuals) and p columns (numeric variables)}

\item{ncp}{an integer, the number of components to keep (value set by default)}

\item{scale.unit}{a boolean, if TRUE (value set by default) then data are scaled to unit variance}

\item{ind_sup}{a vector indicating the indexes of the supplementary individuals}

\item{quanti_sup}{a vector indicating the indexes of the quantitative supplementary variables}
}
\description{
Return PCA results with FactoMineR's style
}
\examples{
library(FactoMineR)
library(FactoMineR2)
data(decathlon)

X <- decathlon[, -c(11:13)]
res <- facto_pca(X, ncp = 5, ind_sup = 1, quanti_sup = 10)
}
