% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{get_eigen}
\alias{get_eigen}
\alias{get_weighted_eigen}
\title{Compute eigvalues and eigvectors}
\usage{
get_eigen(X)

get_weighted_eigen(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\details{
Standardization depends on what you need to perform factor analysis. We
implemented two types:
\itemize{
\item \code{get_weighted_eigen}: This is the default method in FactoMineR to compute
eigvalues and eigvectors.
\item \code{get_eigen}:
}
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  get_eigen()

}
