% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{get_eigen}
\alias{get_eigen}
\title{Compute eigvalues and eigvectors}
\usage{
get_eigen(X)
}
\arguments{
\item{X}{X_active}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\examples{
library(FactoMineR2)

iris[,-5] |>
 standardize(type = "norm") |>
 get_eig()

}
