% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{get_eigen}
\alias{get_eigen}
\alias{eigvalues}
\alias{eigvectors}
\title{Compute eigvalues and eigvectors}
\usage{
get_eigen(X, weights = NULL)

eigvalues(X, weights = NULL)

eigvectors(X, weights = NULL)
}
\arguments{
\item{X}{X_active}

\item{weights}{an optional row weights (by default, a vector of 1 for uniform row weights); the weights are given only for the active individuals}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  get_eigen()

}
