% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{pca_eigen}
\alias{pca_eigen}
\title{Compute eigvalues and eigvectors}
\usage{
pca_eigen(X)
}
\arguments{
\item{X}{X_active}

\item{weighted_row}{row weights}

\item{weighted_col}{column weights}
}
\description{
Return eigvalues and eigvectors of a matrix
}
\details{
Standardization depends on what you need to perform factor analysis. We
implemented two types:
\itemize{
\item \code{pca_weighted_eigen}: This is the default method in FactoMineR to compute
eigvalues, eigvectors and U matrix.
\item \code{pca_eigen}: This is the standard method to compute eigvalues, eigvectors.
}
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  pca_eigen()
}
