% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_contrib}
\alias{pca_ind_contrib}
\title{Compute individual contributions}
\usage{
pca_ind_contrib(
  ind_coord,
  eigs,
  weighted_row = rep(1, nrow(ind_coord))/nrow(ind_coord)
)
}
\arguments{
\item{eigs}{eigs computed by \code{get_eigen} or \code{get_weighted_eigen}}

\item{weighted_row}{row weights}

\item{ind_coords}{individual coordinates}
}
\description{
Return indivdual contributions for each principal component
}
\details{
If you want to compute the contributions of the individuals to the principal
components, you have to change the weighted_col argument to rep(1, nrow(ind_cos2)).
}
\examples{
library(FactoMineR2)

eigs <- iris[, -5] |>
  standardize_norm() |>
  get_weighted_eigen()

eigs |>
  pca_ind_coords() |>
  pca_ind_contrib(eigs) |>
  head()
}
