% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_contrib}
\alias{pca_ind_contrib}
\title{Compute individual contributions}
\usage{
pca_ind_contrib(ind_coords, eigenvalues)
}
\arguments{
\item{ind_coords}{individual coordinates}

\item{eigenvalues}{eigenvalues}
}
\description{
Return indivdual contributions for each principal component
}
\examples{
library(FactoMineR2)

X_active <- iris[, -5] |>
  standardize_norm()

eigs <- X_active |>
  get_eigen()

X_active |>
  pca_ind_coords(eigs$vectors) |>
  pca_ind_contrib(eigs$values) |>
  head()
}
