% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_coords}
\alias{pca_ind_coords}
\title{Compute coordinates for individuals}
\usage{
pca_ind_coords(X, eigvectors)
}
\arguments{
\item{X}{X_active}

\item{eigvectors}{eigenvectors}
}
\description{
Return principal component for individuals
}
\examples{
library(FactoMineR2)

X_active <- iris[,-5] |>
  standardize(type = "norm")

vectors <- X_active |>
  eigvectors()

X_active |>
  pca_ind_coords(vectors)

}
