% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_cos2}
\alias{pca_ind_cos2}
\title{Compute individual squared cosines}
\usage{
pca_ind_cos2(ind_coords)
}
\arguments{
\item{ind_coords}{individual coordinates}
}
\description{
Return indivdual squared cosines for each principal component
}
\examples{
library(FactoMineR2)

X_active <- iris[, -5] |>
  standardize_norm()

eigs <- X_active |>
  get_eigen()

X_active |>
  pca_ind_coords(eigs$vectors) |>
  pca_ind_cos2()
}
