% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ind_cos2}
\alias{pca_ind_cos2}
\title{Compute individual squared cosines}
\usage{
pca_ind_cos2(ind_coords)
}
\arguments{
\item{ind_coords}{individual coordinates}
}
\description{
Return indivdual squared cosines for each principal component
}
\examples{
library(FactoMineR2)

X_active <- iris[,-5] |>
  standardize(type = "norm")

vectors <- X_active |>
  eigvectors()

X_active |>
  pca_ind_coords(vectors) |>
  pca_ind_cos2()

}
