% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_contrib}
\alias{pca_var_contrib}
\title{Compute variable contributions}
\usage{
pca_var_contrib(var_cos2, eigs = NULL, weighted_col = rep(1, ncol(var_cos2)))
}
\arguments{
\item{var_cos2}{variable coordinates}
}
\description{
Return variable contributions
}
\examples{
library(FactoMineR2)

eigs <- iris[, -5] |>
  standardize_norm() |>
  get_weighted_eigen()

eigs |>
  pca_var_coords() |>
  pca_var_cos2() |>
  pca_var_contrib(eigs) |>
  head()
}
