% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_contrib}
\alias{pca_var_contrib}
\title{Compute variable contributions}
\usage{
pca_var_contrib(var_coords)
}
\arguments{
\item{var_coords}{variable coordinates}
}
\description{
Return variable contributions
}
\examples{
library(FactoMineR2)

X_active <- iris[, -5] |>
  standardize_norm()

eigs <- X_active |>
  get_eigen()

eigs$values |>
  pca_var_coords(eigs$vectors) |>
  pca_var_contrib()
}
