% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_coords}
\alias{pca_var_coords}
\title{Compute variable coordinates}
\usage{
pca_var_coords(eigs)
}
\arguments{
\item{eigs}{eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}}
}
\description{
Return variable coordinates
}
\examples{
library(FactoMineR2)

iris[, -5] |>
  pca_standardize_norm() |>
  pca_weighted_eigen() |>
  pca_var_coords() |>
  head()
}
