% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_var_cor}
\alias{pca_var_cor}
\title{Compute variable correlation}
\usage{
pca_var_cor(eigs)
}
\arguments{
\item{eigs}{eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}}
}
\value{
A dataframe of variable correlation.
}
\description{
Return variable correlation
}
\examples{
library(booklet)

iris[, -5] |>
  pca_standardize_norm() |>
  pca_weighted_eigen() |>
  pca_var_cor() |>
  head()
}
